pyproject.toml
[project]
name = "acucys-ctf"
version = "0.1.0"
description = "Official Discord bot for the ACUCyS CTF event"
authors = [{ name = "PyPylia", email = "me@pypylia.dev" }]
license = { text = "GPLv3" }
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "aiohttp (>=3.13.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "typing-json (>=0.1.3,<0.2.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
acucys-ctf = "acucys_ctf.__main__:main"
